requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
torch==2.0.1
tqdm==4.65.0
transformers==4.27.2
